+++ lib/commands/create_command/create_command.dart
@@ -1,0 +1,3 @@
+abstract class CreateCommand {
+  Future<void> execute();
+}