+++ lib/commands/init_command/boc_init_command.dart
@@ -101,6 +101,10 @@ class BlocInitCommand extends Command with Generator {
         content: InitGenerator.networkImageWidgetFileContent.replaceAppName,
       ),
       writeFile(
+        path: Constants.searchFieldWidgetPath.actualPath(),
+        content: InitGenerator.searchFieldWidgetFileContent.replaceAppName,
+      ),
+      writeFile(
         path: Constants.textStyleFilePath.actualPath(),
         content: InitGenerator.textStyleFileContent.replaceAppName,
       ),