+++ lib/flutter_bloc_cli.dart
@@ -9,8 +9,7 @@ class CliCommand {
   CliCommand._();
 
   Command findCommand() {
-    final Map<String, dynamic> commandData = CliDataProvider
-        .instance.commandData[CliDataProvider.instance.codePattern.value];
+    final Map<String, dynamic> commandData = CliDataProvider.instance.commandData[CliDataProvider.instance.codePattern.value];
     List<String> args = CliDataProvider.instance.args;
     if (args.isEmpty) {
       return commandData["help"]["command"];
@@ -18,23 +17,26 @@ class CliCommand {
     if (!commandData.containsKey(args.first)) {
       return CommandNotFound();
     }
-    if (commandData[args.first]["args_length"] != args.length) {
+    if (commandData[args.first]["args_length"] != null && commandData[args.first]["args_length"] != args.length) {
       return InvalidCommand();
     }
     if (!commandData[args.first]["has_sub_command"]) {
       return commandData[args.first]["command"];
     }
-    return _validate(commandData[args.first], args) ??
-        commandData["help"]["command"];
+    return _validate(commandData[args.first], args) ?? commandData["help"]["command"];
   }
 
   Command? _validate(Map<String, dynamic> data, List<String> args) {
-    if (data['args_length'] != args.length) {
+    if (data['sub_commands'] is! Map) {
       return InvalidCommand();
     }
-    if ((data['sub_commands'] ?? []).contains(args[1])) {
-      return data['command'];
+    int? argsLen = data['sub_commands']['args_length'];
+    if (argsLen != null && argsLen != args.length) {
+      return InvalidCommand();
     }
-    return null;
+    if (!(data['sub_commands'] as Map).containsKey(args[1])) {
+      return InvalidCommand();
+    }
+    return data['sub_commands'][args[1]]['command'];
   }
 }