+++ CHANGELOG.md
@@ -9,3 +9,7 @@
 ### 1.0.2 - File path updated 
 
 - Updated file path separator for Mac and Linux
+
+### 1.0.3 - Added screen creation command.
+
+- Added a feature for creating multiple screens at once.